% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{cblasr}
\alias{cblasr}
\alias{cblasr-package}
\title{cblasr}
\description{
Provide the 'cblas.h' header file for R. 'CBLAS' is a
collection of wrappers that provide a C interface to the FORTRAN 'BLAS'
library. Currently only the double precision / double complex 'BLAS'
routines are supported.
}
\author{
Yi Pan
}
