% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example_cblas_dgemm}
\alias{example_cblas_dgemm}
\title{Example: Matrix multiplication using cblas_dgemm}
\usage{
example_cblas_dgemm()
}
\description{
Matrix multiplication using cblas_dgemm
}
\examples{
## Expected output:
## [ 367.76, 368.12
##   674.06, 674.72 ]
example_cblas_dgemm()
}
